wegodang/requirements.txt
bcrypt==3.2.2
Django==4.0.5
django-cors-headers==3.13.0
mysqlclient==2.1.1
PyJWT==2.4.0
PyMySQL==1.0.2
requests==2.28.1
sqlparse==0.4.2